% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_home.R
\name{get_home}
\alias{get_home}
\title{get_home}
\usage{
get_home()
}
\value{
A character vector containing the path to the Home directory
}
\description{
Finds the path to the home directory on any OS
}
\examples{
my_home <- get_home()
}
