% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reload.R
\name{reload}
\alias{reload}
\title{reload}
\usage{
reload(clear_env = TRUE)
}
\arguments{
\item{clear_env}{Set to FALSE if you don't want the environment cleared. Defaults to TRUE}

\item{use_rcpp}{Set to FALSE if you're not using Rcpp. Defaults to TRUE}
}
\description{
Clears the environment and reloads the current package
}
\examples{
reload()
reload(clear_env = FALSE)
reload(use_rcpp = FALSE)
reload(clear_env = FALSE, use_rcpp = FALSE)
}
