% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_reload.R
\name{setup_reload}
\alias{setup_reload}
\title{setup_reload}
\usage{
setup_reload(package_dir, working_dir)
}
\arguments{
\item{package_dir}{A string representing the base directory of the R package in development}

\item{working_dir}{A string representing the directory you are working in}
}
\description{
creates a settings file in the users home directory that will be used by reload()
}
\examples{
setup_reload(package_dir = "C:/packages/my_package", working_dir = "C:/Analysis/analysis_23") 
setup_reload(package_dir = file.path(get_home(), "Documents", "Packages", "my_package"), working_dir = file.path(get_desktop(), "my_working_directory"))
}
